[app名称,图标(使用picasso加载),详情,星级平分的显示]

+++ app/src/main/java/com/itheima/googleplay25/beans/HomeBean.java
@@ -33,7 +33,7 @@ public class HomeBean {
         public String name;
         public String packageName;
         public String iconUrl;
-        public int    stars;
+        public float    stars;
         public int    size;
         public String downloadUrl;
         public String des;
